fix(tile-converter): i3s->3dtiles regression
