--- modules/tile-converter/src/workers/i3s-attributes-worker.ts
@@ -1,7 +1,0 @@
-import {createWorker} from '@loaders.gl/worker-utils';
-import {convertAttributes} from '../i3s-converter/helpers/geometry-converter';
-
-createWorker(
-  async (data, options = {}) =>
-    await convertAttributes(data, options.materialAndTextureList, options.useCartesianPositions)
-);
